Refactor job persistence for predictive reretryability
