+++ control-plane/drizzle/0035_eminent_fantastic_four.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "jobs" ADD COLUMN "predicted_to_be_retryable" boolean;--> statement-breakpoint
+ALTER TABLE "jobs" ADD COLUMN "predicted_to_be_retryable_reason" text;
